mv creating and deleting methods to table classes
